+++ Engine/Graphics/src/Graphics/RHI/RHI_FSR.cpp
@@ -35,7 +35,7 @@ namespace Insight
         {
             Maths::Vector2 render_resolution = RenderGraph::Instance().GetRenderResolution();
             Maths::Vector2 output_resolution = RenderGraph::Instance().GetOutputResolution();
-            CreateContext(render_resolution.x, render_resolution.y, output_resolution.x, output_resolution.y);
+            CreateContext(static_cast<u32>(render_resolution.x), static_cast<u32>(render_resolution.y), static_cast<u32>(output_resolution.x), static_cast<u32>(output_resolution.y));
 
             //Core::EventSystem::Instance().AddEventListener(this, Core::EventType::Graphics_Swapchain_Resize, [this](const Core::Event& event)
             //    {
@@ -61,7 +61,7 @@ namespace Insight
                             RenderContext* render_context = &RenderContext::Instance();
                             Maths::Vector2 render_resolution = RenderGraph::Instance().GetRenderResolution();
                             Maths::Vector2 output_resolution = RenderGraph::Instance().GetOutputResolution();
-                            CreateContext(render_resolution.x, render_resolution.y, output_resolution.x, output_resolution.y);
+                            CreateContext(static_cast<u32>(render_resolution.x), static_cast<u32>(render_resolution.y), static_cast<u32>(output_resolution.x), static_cast<u32>(output_resolution.y));
                         });
                 });
         }
